requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.8.2
Werkzeug==3.0.1
numpy==1.26.4
opencv-python==4.9.0.80
requests==2.31.0
python-dotenv==1.0.1
openai==1.12.0